5_Return_predictions_on_demand/requirements.txt
joblib==1.4.2
numpy==1.26.4
pandas==2.2.2
PyJWT==2.10.1
scikit-learn==1.5.2
requests==2.32.3
boto3==1.35.81
awslambdaric==3.0.0
